+++ src/location/mod.rs
@@ -1,4 +1,5 @@
 use serde::{Deserialize, Serialize};
+use strum_macros::EnumString;
 
 use crate::util;