+++ src/weather/providers/owm/mod.rs
@@ -2,6 +2,7 @@ use log::debug;
 
 use crate::location::Location;
 use crate::util;
+use crate::weather::error::UndefinedCondition;
 use crate::weather::providers::CurrentWeather;
 use crate::weather::providers::owm::response::Response;
 use crate::weather::Weather;
@@ -9,7 +10,9 @@ use crate::weather::weather_condition::WeatherCondition;
 
 mod response;
 
-pub struct OpenWeatherMap {}
+pub struct OpenWeatherMap;
+
+const BASE_URL: &str = "http://api.openweathermap.org/data/2.5";
 
 impl CurrentWeather for OpenWeatherMap {
     fn current_weather(&self, location: &Location, api_key: &str) -> util::Result<Weather> {
@@ -28,15 +31,13 @@ impl CurrentWeather for OpenWeatherMap {
 
 impl OpenWeatherMap {
     pub fn new() -> Self {
-        OpenWeatherMap {}
+        OpenWeatherMap
     }
 
     fn build_url(&self, location: &Location, api_key: &str) -> String {
-        let base_url = "http://api.openweathermap.org/data/2.5";
-
         let url = format!(
             "{}/weather?lat={}&lon={}&APPID={}",
-            base_url,
+            BASE_URL,
             location.lat,
             location.lon,
             api_key
@@ -45,38 +46,23 @@ impl OpenWeatherMap {
         url
     }
 
-    fn parse_weather_condition(&self, response: &Response) -> Result<WeatherCondition, String> {
+    fn parse_weather_condition(&self, response: &Response) -> util::Result<WeatherCondition> {
         let id = response.weather[0].id;
-        let first_digit = id.to_string()[0..1].parse::<i32>().unwrap();
+        let first_digit = id.to_string()[0..1].parse::<i32>()?;
 
         match first_digit {
             2 => Ok(WeatherCondition::Thunderstorm),
 
             // rain
             3 => match id {
-                300 => Ok(WeatherCondition::Rain),
-                301 => Ok(WeatherCondition::Rain),
-                302 => Ok(WeatherCondition::HeavyRain),
-                310 => Ok(WeatherCondition::Rain),
-                311 => Ok(WeatherCondition::HeavyRain),
-                312 => Ok(WeatherCondition::HeavyRain),
-                313 => Ok(WeatherCondition::HeavyRain),
-                314 => Ok(WeatherCondition::HeavyRain),
-                321 => Ok(WeatherCondition::HeavyRain),
-                _ => Err("Undefined weather condition".to_string())
+                300 | 301 | 310 => Ok(WeatherCondition::Rain),
+                302 | 311 | 312 | 313 | 314 | 321 => Ok(WeatherCondition::HeavyRain),
+                _ => Err(UndefinedCondition.into())
             },
             5 => match id {
-                500 => Ok(WeatherCondition::Rain),
-                501 => Ok(WeatherCondition::HeavyRain),
-                502 => Ok(WeatherCondition::HeavyRain),
-                503 => Ok(WeatherCondition::HeavyRain),
-                504 => Ok(WeatherCondition::HeavyRain),
-                511 => Ok(WeatherCondition::HeavyRain),
-                520 => Ok(WeatherCondition::Rain),
-                521 => Ok(WeatherCondition::HeavyRain),
-                522 => Ok(WeatherCondition::HeavyRain),
-                531 => Ok(WeatherCondition::HeavyRain),
-                _ => Err("Undefined weather condition".to_string())
+                500 | 520 => Ok(WeatherCondition::Rain),
+                501 | 502 | 503 | 504 | 511 | 521 | 522 | 531 => Ok(WeatherCondition::HeavyRain),
+                _ => Err(UndefinedCondition.into())
             },
 
             6 => Ok(WeatherCondition::Snow),
@@ -87,12 +73,11 @@ impl OpenWeatherMap {
                 800 => Ok(WeatherCondition::ClearSky),
                 801 => Ok(WeatherCondition::FewClouds),
                 802 => Ok(WeatherCondition::Clouds),
-                803 => Ok(WeatherCondition::ManyClouds),
-                804 => Ok(WeatherCondition::ManyClouds),
-                _ => Err("Undefined weather condition".to_string())
+                803 | 804 => Ok(WeatherCondition::ManyClouds),
+                _ => Err(UndefinedCondition.into())
             },
 
-            _ => Err("Undefined weather condition".to_string())
+            _ => Err(UndefinedCondition.into())
         }
     }
 }