+++ src/weather/weather_condition.rs
@@ -1,8 +1,10 @@
 use std::collections::HashMap;
 
 use serde::{Deserialize, Serialize};
+use strum_macros::Display;
 
 #[derive(Debug, Hash, Eq, PartialEq, Serialize, Deserialize, Display)]
+#[strum(serialize_all = "snake_case")]
 pub enum WeatherCondition {
     ClearSky,
     FewClouds,
@@ -19,16 +21,18 @@ pub enum WeatherCondition {
 pub fn default_icons() -> HashMap<String, String> {
     let mut icons = HashMap::new();
 
-    icons.insert(WeatherCondition::ClearSky.to_string(), "".to_string());
-    icons.insert(WeatherCondition::FewClouds.to_string(), "".to_string());
-    icons.insert(WeatherCondition::Clouds.to_string(), "摒".to_string());
-    icons.insert(WeatherCondition::ManyClouds.to_string(), "".to_string());
-    icons.insert(WeatherCondition::RainSun.to_string(), "".to_string());
-    icons.insert(WeatherCondition::Rain.to_string(), "".to_string());
-    icons.insert(WeatherCondition::HeavyRain.to_string(), "歹".to_string());
-    icons.insert(WeatherCondition::Thunderstorm.to_string(), "".to_string());
-    icons.insert(WeatherCondition::Snow.to_string(), "".to_string());
-    icons.insert(WeatherCondition::Mist.to_string(), "".to_string());
+    icons.insert(WeatherCondition::ClearSky, "");
+    icons.insert(WeatherCondition::FewClouds, "");
+    icons.insert(WeatherCondition::Clouds, "摒");
+    icons.insert(WeatherCondition::ManyClouds, "");
+    icons.insert(WeatherCondition::RainSun, "");
+    icons.insert(WeatherCondition::Rain, "");
+    icons.insert(WeatherCondition::HeavyRain, "歹");
+    icons.insert(WeatherCondition::Thunderstorm, "");
+    icons.insert(WeatherCondition::Snow, "");
+    icons.insert(WeatherCondition::Mist, "");
 
-    icons
+    icons.into_iter()
+        .map(|(key, value)| (key.to_string(), value.to_string()))
+        .collect()
 }