+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- BREAKING: Rename icon names in config to snake case (e.g. ClearSky &rarr; clear_sky)
 - Add polybar config to open weather forecast in browser on module click
 - Add option to print current city
 - Improve debug output